Address more PR comments, rename module names to adhere to guidelines
