+++ src/controller/python/BUILD.gn
@@ -226,9 +226,10 @@ chip_python_wheel_action("chip-core") {
         "chip/storage/__init__.py",
         "chip/utils/CommissioningBuildingBlocks.py",
         "chip/utils/__init__.py",
-        "chip/yaml/DataModelLookup.py",
-        "chip/yaml/YamlParser.py",
-        "chip/yaml/YamlUtils.py",
+        "chip/yaml/data_model_lookup.py",
+        "chip/yaml/format_converter.py",
+        "chip/yaml/errors.py",
+        "chip/yaml/parser.py",
         "chip/yaml/__init__.py",
       ]
